general cleanups
bench: 10513080

+++ src/nnue/scalar.rs
@@ -24,7 +24,7 @@ pub(super) fn infer(stm: &[i16; HL_SIZE], nstm: &[i16; HL_SIZE]) -> i32 {
     *left = crelu(stm);
     *right = crelu(nstm);
 
-    let mut result = NETWORK.l1.bias[0] as i32;
+    let mut result = NETWORK.l1.bias[0];
 
     for i in 0..activated.len() {
         result += activated[i] as i32 * activated[i] as i32 * NETWORK.l1.w[0][i] as i32;
